FileAPI/new-Blob.html

<!doctype html>
<html>
 <head>
  <title>Blob constructor</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id="log"></div>
  <script>
    test(function(){
        assert_false('toNativeLineEndings' in window);
    }, document.title+', non-support for obsolete "toNativeLineEndings" thing');

    var t_slice = async_test(document.title+', slice');
    t_slice.step(function(){
        var view = new Uint8Array([0, 255, 0]);
        var b1 = new Blob(['abcd']);
        var b2 = new Blob([view.buffer, b1, 'efgh', 'ijklmnopqrstuvwxyz']);
        var b3 = b2.slice(1, 4);
        b3.expectedText = '\uFFFD\u0000a';
        b3.expectedSize = 3;
        b3.msg = 'b3';
        var b4 = b2.slice(4, 8);
        b4.expectedText = 'bcde';
        b4.expectedSize = 4;
        b4.msg = 'b4';
        var b5 = b2.slice(8, 12);
        b5.expectedText = 'fghi';
        b5.expectedSize = 4;
        b5.msg = 'b5';
        var b6 = b2.slice(1, 12);
        b6.expectedText = '\uFFFD\u0000abcdefghi';
        b6.expectedSize = 11;
        b6.msg = 'b6';
        [b3, b4, b5, b6].forEach(this.step_func(function(b) {
            assert_equals(b.size, b.expectedSize, b.msg);
            var reader = new FileReader();
            reader.onload = this.step_func(function(e) {
                assert_equals(reader.result, b.expectedText, b.msg);
                if (b === b6)
                    this.done();
            });
            reader.readAsText(b, 'utf-8');
        }));
    });
  </script>
 </body>
</html>
